Added composer and gradio  App UI
